bibliotheek/1728469813381cecfb480-32a0-4e89-a564-3eeafbf8765f/requirements.txt
Flask~=3.0.3
redis~=5.0.8
python-dotenv~=1.0.1
SQLAlchemy~=2.0.32
flask_sqlalchemy
flask_bcrypt
pymysql
gunicorn